+++ vorx/core/__init__.py
@@ -1,5 +1,0 @@
-from maths import vectors
-
-v = vectors.Vector2(1, 2)
-
-print(v)